Separate out apis for varaibles expression into ExpressionEvaluator
