+++ datacapture/src/main/java/com/google/android/fhir/datacapture/QuestionnaireViewModel.kt
@@ -25,7 +25,6 @@ import ca.uhn.fhir.context.FhirContext
 import ca.uhn.fhir.context.FhirVersionEnum
 import ca.uhn.fhir.parser.IParser
 import com.google.android.fhir.datacapture.enablement.EnablementEvaluator
-import com.google.android.fhir.datacapture.fhirpath.FHIRPathEngineHostServices
 import com.google.android.fhir.datacapture.validation.QuestionnaireResponseItemValidator
 import com.google.android.fhir.datacapture.validation.QuestionnaireResponseValidator.checkQuestionnaireResponse
 import com.google.android.fhir.datacapture.validation.ValidationResult
@@ -36,17 +35,11 @@ import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
-import org.hl7.fhir.exceptions.FHIRException
-import org.hl7.fhir.r4.hapi.ctx.HapiWorkerContext
-import org.hl7.fhir.r4.model.Base
-import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
-import org.hl7.fhir.r4.model.Expression
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.ResourceType
 import org.hl7.fhir.r4.model.ValueSet
-import org.hl7.fhir.r4.utils.FHIRPathEngine
 import timber.log.Timber
 
 internal class QuestionnaireViewModel(application: Application, state: SavedStateHandle) :
@@ -205,13 +198,6 @@ internal class QuestionnaireViewModel(application: Application, state: SavedStat
   /** Tracks modifications in order to update the UI. */
   private val modificationCount = MutableStateFlow(0)
 
-  private val fhirPathEngine: FHIRPathEngine =
-    with(FhirContext.forCached(FhirVersionEnum.R4)) {
-      FHIRPathEngine(HapiWorkerContext(this, this.validationSupport)).apply {
-        hostServices = FHIRPathEngineHostServices
-      }
-    }
-
   /**
    * Contains [QuestionnaireResponse.QuestionnaireResponseItemComponent]s that have been modified by
    * the user. [QuestionnaireResponse.QuestionnaireResponseItemComponent]s that have not been
@@ -248,6 +234,10 @@ internal class QuestionnaireViewModel(application: Application, state: SavedStat
     if (questionnaireItem.hasNestedItemsWithinAnswers) {
       questionnaireResponseItem.addNestedItemsToAnswer(questionnaireItem)
     }
+
+    //    val result =
+    // ExpressionEvaluator.evaluateQuestionnaireItemVariableExpression(questionnaireItem.variableExpressions.last(), questionnaire, questionnaireResponse, questionnaireItemParentMap, questionnaireItem)
+    //    println(result)
 
     modifiedQuestionnaireResponseItemSet.add(questionnaireResponseItem)
     modificationCount.update { it + 1 }
@@ -576,173 +566,6 @@ internal class QuestionnaireViewModel(application: Application, state: SavedStat
 
     return null
   }
-
-  /**
-   * Evaluates variable expression and return the evaluated result. To evaluate root [Questionnaire]
-   * level variable expressions, we pass only an expression to this function and for variables
-   * expressions defined at questionnaire item [Questionnaire.QuestionnaireItemComponent] level, we
-   * pass expression and respective questionnaire item to this function
-   *
-   * If an expression is simple and evaluates on first evaluation, the function returns the
-   * evaluated value otherwise we parse the expression using regex [Regex] for variable (For
-   * example: A variable name could be %weight) and build a list of variables that the expression
-   * contains and for every variable, we first find it at origin, then up in the parent hierarchy
-   * and then at root/questionnaire level, if found we get their expressions and pass them into the
-   * same function to evaluate its value recursively, we put the variable name and its evaluated
-   * value into the map [Map] to use this map to pass into fhirPathEngine's evaluate method to apply
-   * the evaluated values to the expression being evaluated.
-   *
-   * @param expression the [Expression] Variable expression
-   * @param origin the [Questionnaire.QuestionnaireItemComponent] where this expression is defined,
-   * null if expression defined at questionnaire [Questionnaire] level
-   *
-   * @return [Base] the result of expression
-   */
-  internal fun evaluateExpression(
-    expression: Expression,
-    origin: Questionnaire.QuestionnaireItemComponent? = null
-  ): Base? {
-
-    val result = evaluateVariable(expression)
-    if (result != null) {
-      return result
-    }
-
-    buildMap<String, Base?> {
-      buildList {
-        val variableMatches = variableRegex.findAll(expression.expression)
-
-        addAll(
-          variableMatches.map { it.groupValues[1] }.toList().filterNot { variable ->
-            reservedVariables.any { it == variable }
-          }
-        )
-      }
-        .takeIf { it.isNotEmpty() }
-        ?.forEach { variableName ->
-          findVariable(variableName, origin)?.also { (questionnaireItem, expression) ->
-            put(expression.name, evaluateExpression(expression, questionnaireItem))
-          }
-        }
-    }
-      .also {
-        return evaluateVariable(expression, it)
-      }
-  }
-
-  /**
-   * Finds a variable, first checks at origin if not found then checks in parent hierarchy, if not
-   * found then checks at root level and return the Pair
-   *
-   * @param variableName the [String] to match the variable
-   * @param origin the [Questionnaire.QuestionnaireItemComponent] from where we have to track
-   * hierarchy
-   *
-   * @return [Pair] containing [Questionnaire.QuestionnaireItemComponent] and [Expression]
-   */
-  private fun findVariable(
-    variableName: String,
-    origin: Questionnaire.QuestionnaireItemComponent?
-  ): Pair<Questionnaire.QuestionnaireItemComponent?, Expression>? {
-    return findVariableAtOrigin(variableName, origin)
-      ?: findVariableInParent(variableName, origin) ?: findVariableAtRoot(variableName)
-  }
-
-  /**
-   * Finds the specific variable name [String] at the origin
-   * [Questionnaire.QuestionnaireItemComponent]
-   *
-   * @param origin the [Questionnaire.QuestionnaireItemComponent] from where we have to track
-   * hierarchy up in the parent
-   * @param variableName the [String] to match the variable in the parent hierarchy
-   *
-   * @return [Pair] containing [Questionnaire.QuestionnaireItemComponent] and [Expression]
-   */
-  private fun findVariableAtOrigin(
-    variableName: String,
-    origin: Questionnaire.QuestionnaireItemComponent? = null
-  ): Pair<Questionnaire.QuestionnaireItemComponent?, Expression>? {
-    origin?.variableExpressions?.find { it.name == variableName }.also {
-      it?.let {
-        return Pair(origin, it)
-      }
-    }
-    return null
-  }
-
-  /**
-   * Finds the specific variable name [String] in the parent hierarchy of origin
-   * [Questionnaire.QuestionnaireItemComponent]
-   *
-   * @param origin the [Questionnaire.QuestionnaireItemComponent] from where we have to track
-   * hierarchy up in the parent
-   * @param variableName the [String] to match the variable in the parent hierarchy
-   *
-   * @return [Pair] containing [Questionnaire.QuestionnaireItemComponent] and [Expression]
-   */
-  private fun findVariableInParent(
-    variableName: String,
-    origin: Questionnaire.QuestionnaireItemComponent? = null
-  ): Pair<Questionnaire.QuestionnaireItemComponent?, Expression>? {
-    var parent = questionnaireItemParentMap[origin]
-    while (parent != null) {
-      parent.variableExpressions.find { it.name == variableName }.also {
-        it?.let {
-          return Pair(parent!!, it)
-        }
-      }
-      parent = questionnaireItemParentMap[parent]
-    }
-    return null
-  }
-
-  /**
-   * Finds the specific variable name [String] at root/questionnaire [Questionnaire] level
-   *
-   * @param variableName the [String] to match the variable at questionnaire [Questionnaire] level
-   *
-   * @return [Expression] the matching expression
-   */
-  private fun findVariableAtRoot(
-    variableName: String
-  ): Pair<Questionnaire.QuestionnaireItemComponent?, Expression>? {
-    questionnaire.variableExpressions.find { it.name == variableName }.also {
-      it?.let {
-        return Pair(null, it)
-      }
-    }
-    return null
-  }
-
-  /**
-   * Evaluates the value of variable expression and return the evaluated value
-   *
-   * @param expression the [Expression] the expression to evaluate
-   * @param inputVariables the [Map] of Variable names to their values
-   *
-   * @return [Base] the result of expression
-   */
-  private fun evaluateVariable(
-    expression: Expression,
-    inputVariables: Map<String, Base?> = mapOf()
-  ) =
-    try {
-
-      require(!expression.name.isNullOrEmpty()) {
-        "Expression name should be a valid expression name"
-      }
-
-      require(expression.hasLanguage() && expression.language == "text/fhirpath") {
-        "Unsupported expression language, language should be text/fhirpath"
-      }
-
-      fhirPathEngine
-        .evaluate(inputVariables, questionnaireResponse, null, null, expression.expression)
-        .firstOrNull()
-    } catch (exception: FHIRException) {
-      Timber.w("Could not evaluate expression with FHIRPathEngine", exception)
-      null
-    }
 }
 
 /** Questionnaire state for the Fragment to consume. */
@@ -778,8 +601,3 @@ internal fun QuestionnairePagination.nextPage(): QuestionnairePagination {
 
 internal val QuestionnairePagination.hasNextPage: Boolean
   get() = pages.any { it.index > currentPageIndex && it.enabled }
-
-private val reservedVariables =
-  listOf("sct", "loinc", "ucum", "resource", "rootResource", "context", "map-codes")
-
-private val variableRegex = Regex("[%]([A-Za-z0-9\\-]{1,64})")
