+++ datacapture/src/test/java/com/google/android/fhir/datacapture/QuestionnaireViewModelTest.kt
@@ -27,7 +27,6 @@ import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.EXTRA
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.EXTRA_QUESTIONNAIRE_JSON_URI
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.EXTRA_QUESTIONNAIRE_RESPONSE_JSON_STRING
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.EXTRA_QUESTIONNAIRE_RESPONSE_JSON_URI
-import com.google.android.fhir.datacapture.common.datatype.asStringValue
 import com.google.android.fhir.datacapture.testing.DataCaptureTestApplication
 import com.google.android.fhir.datacapture.validation.ValidationResult
 import com.google.android.fhir.datacapture.views.QuestionnaireItemViewItem
@@ -43,15 +42,12 @@ import org.hl7.fhir.instance.model.api.IBaseResource
 import org.hl7.fhir.r4.model.BooleanType
 import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
-import org.hl7.fhir.r4.model.Expression
 import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.IntegerType
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.StringType
-import org.hl7.fhir.r4.model.Type
 import org.hl7.fhir.r4.model.ValueSet
-import org.junit.Assert
 import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
@@ -1686,353 +1682,6 @@ class QuestionnaireViewModelTest(
 
     assertThat(viewModel.getQuestionnaireItemViewItemList().last().questionnaireItem.linkId)
       .isEqualTo("parent-question")
-  }
-
-  @Test
-  fun questionnaireRootLevel_simpleVariableExpression_shouldReturnNotNullValue() = runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "A"
-              language = "text/fhirpath"
-              expression = "1"
-            }
-          )
-        }
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result = viewModel.evaluateExpression(questionnaire.variableExpressions.first())
-
-    assertThat((result as Type).asStringValue()).isEqualTo("1")
-  }
-
-  @Test
-  fun questionnaireRootLevel_variableDependOnOneOtherVariable_shouldReturnNotNullValue() =
-      runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "A"
-              language = "text/fhirpath"
-              expression = "1"
-            }
-          )
-        }
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "B"
-              language = "text/fhirpath"
-              expression = "%A + 1"
-            }
-          )
-        }
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result = viewModel.evaluateExpression(questionnaire.variableExpressions.last())
-
-    assertThat((result as Type).asStringValue()).isEqualTo("2")
-  }
-
-  @Test
-  fun questionnaireItemLevel_variableDependOnOneOtherVariableInParent_shouldReturnNotNullValue() =
-      runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        id = "a-questionnaire"
-        addItem(
-          Questionnaire.QuestionnaireItemComponent().apply {
-            linkId = "a-group-item"
-            text = "a question"
-            type = Questionnaire.QuestionnaireItemType.GROUP
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "A"
-                  language = "text/fhirpath"
-                  expression = "1"
-                }
-              )
-            }
-            addItem(
-              Questionnaire.QuestionnaireItemComponent().apply {
-                linkId = "an-item"
-                text = "a question"
-                type = Questionnaire.QuestionnaireItemType.TEXT
-                addExtension().apply {
-                  url = EXTENSION_VARIABLE_URL
-                  setValue(
-                    Expression().apply {
-                      name = "B"
-                      language = "text/fhirpath"
-                      expression = "%A + 1"
-                    }
-                  )
-                }
-              }
-            )
-          }
-        )
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result =
-      viewModel.evaluateExpression(
-        viewModel.questionnaire.item[0].item[0].variableExpressions.last(),
-        viewModel.questionnaire.item[0].item[0]
-      )
-
-    assertThat((result as Type).asStringValue()).isEqualTo("2")
-  }
-
-  @Test
-  fun questionnaireRootLevel_variableDependOnMultipleOtherVariable_shouldReturnNotNullValue() =
-      runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "A"
-              language = "text/fhirpath"
-              expression = "1"
-            }
-          )
-        }
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "B"
-              language = "text/fhirpath"
-              expression = "2"
-            }
-          )
-        }
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "C"
-              language = "text/fhirpath"
-              expression = "%A + %B"
-            }
-          )
-        }
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result = viewModel.evaluateExpression(questionnaire.variableExpressions.last())
-
-    assertThat((result as Type).asStringValue()).isEqualTo("3")
-  }
-
-  @Test
-  fun questionnaireRootLevel_variableDependOnMissingVariable_shouldReturnNull() = runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addExtension().apply {
-          url = EXTENSION_VARIABLE_URL
-          setValue(
-            Expression().apply {
-              name = "A"
-              language = "text/fhirpath"
-              expression = "%B + 1"
-            }
-          )
-        }
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result = viewModel.evaluateExpression(questionnaire.variableExpressions.last())
-
-    assertThat(result).isEqualTo(null)
-  }
-
-  @Test
-  fun questionnaireItemLevel_variableDependOnOneOtherVariableAtOrigin_shouldReturnNotNullValue() =
-      runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addItem(
-          Questionnaire.QuestionnaireItemComponent().apply {
-            linkId = "an-item"
-            text = "a question"
-            type = Questionnaire.QuestionnaireItemType.TEXT
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "B"
-                  language = "text/fhirpath"
-                  expression = "1"
-                }
-              )
-            }
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "A"
-                  language = "text/fhirpath"
-                  expression = "%B + 1"
-                }
-              )
-            }
-          }
-        )
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result =
-      viewModel.evaluateExpression(
-        questionnaire.item[0].variableExpressions.last(),
-        questionnaire.item[0]
-      )
-
-    assertThat((result as Type).asStringValue()).isEqualTo("2")
-  }
-
-  @Test
-  fun questionnaireItemLevel_variableDependOnMissingVariableAtOrigin_shouldReturnNull() =
-      runBlocking {
-    val questionnaire =
-      Questionnaire().apply {
-        addItem(
-          Questionnaire.QuestionnaireItemComponent().apply {
-            linkId = "an-item"
-            text = "a question"
-            type = Questionnaire.QuestionnaireItemType.TEXT
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "A"
-                  language = "text/fhirpath"
-                  expression = "%B + 1"
-                }
-              )
-            }
-          }
-        )
-      }
-
-    val viewModel = createQuestionnaireViewModel(questionnaire)
-    val result =
-      viewModel.evaluateExpression(
-        questionnaire.item[0].variableExpressions.last(),
-        questionnaire.item[0]
-      )
-
-    assertThat(result).isEqualTo(null)
-  }
-
-  @Test
-  fun questionnaireVariables_missingExpressionName_shouldThrowIllegalArgumentException() {
-    Assert.assertThrows(IllegalArgumentException::class.java) {
-      runBlocking {
-        val questionnaire =
-          Questionnaire().apply {
-            id = "a-questionnaire"
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  language = "text/fhirpath"
-                  expression = "%resource.repeat(item).where(linkId='an-item').answer.first().value"
-                }
-              )
-            }
-          }
-
-        val viewModel = createQuestionnaireViewModel(questionnaire)
-        viewModel.evaluateExpression(viewModel.questionnaire.variableExpressions.first())
-      }
-    }
-  }
-
-  @Test
-  fun questionnaireVariables_missingExpressionLanguage_shouldThrowIllegalArgumentException() {
-    Assert.assertThrows(IllegalArgumentException::class.java) {
-      runBlocking {
-        val questionnaire =
-          Questionnaire().apply {
-            id = "a-questionnaire"
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "X"
-                  expression = "1"
-                }
-              )
-            }
-          }
-
-        val viewModel = createQuestionnaireViewModel(questionnaire)
-        viewModel.evaluateExpression(viewModel.questionnaire.variableExpressions.first())
-      }
-    }
-  }
-
-  @Test
-  fun questionnaireVariables_unSupportedExpressionLanguage_shouldThrowIllegalArgumentException() {
-    Assert.assertThrows(IllegalArgumentException::class.java) {
-      runBlocking {
-        val questionnaire =
-          Questionnaire().apply {
-            id = "a-questionnaire"
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "X"
-                  expression = "1"
-                  language = "application/x-fhir-query"
-                }
-              )
-            }
-          }
-
-        val viewModel = createQuestionnaireViewModel(questionnaire)
-        viewModel.evaluateExpression(viewModel.questionnaire.variableExpressions.first())
-      }
-    }
-  }
-
-  @Test
-  fun questionnaireVariables_missingExpression_shouldThrowNullPointerException() {
-    Assert.assertThrows(NullPointerException::class.java) {
-      runBlocking {
-        val questionnaire =
-          Questionnaire().apply {
-            id = "a-questionnaire"
-            addExtension().apply {
-              url = EXTENSION_VARIABLE_URL
-              setValue(
-                Expression().apply {
-                  name = "X"
-                  language = "text/fhirpath"
-                }
-              )
-            }
-          }
-
-        val viewModel = createQuestionnaireViewModel(questionnaire)
-        viewModel.evaluateExpression(viewModel.questionnaire.variableExpressions.first())
-      }
-    }
   }
 
   private fun createQuestionnaireViewModel(
